+++ Player/Player/doc/Doxygen/html/search/all_3.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['display_5fattributes',['display_attributes',['../class_a_t_t_r_i_b_u_t_e_s.html#ac1653b73b37c39c05aa08abb4885d579',1,'ATTRIBUTES']]],
+  ['display_5fcharacter',['display_character',['../class_c_h_a_r_a_c_t_e_r.html#a49b688dd97468943d50a34c68b1a21e9',1,'CHARACTER']]]
+];