+++ Player/Player/doc/Doxygen/html/search/defines_0.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['_5fcharacter_5fh',['_CHARACTER_H',['../_c_h_a_r_a_c_t_e_r_8h.html#a20ba65d312300f6c2a255b39f10b7af3',1,'CHARACTER.h']]],
+  ['_5fmacro_5fh',['_MACRO_H',['../_m_a_c_r_o_8h.html#ad43c564e4012666282d44e9c4ec3af5a',1,'MACRO.h']]]
+];