+++ Player/Player/doc/CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-##[0.1.0] - 2019-11/24
+##[0.2.0] - 2019-11-24
+### Added
+- Added MACRO.h, ATTRIBUTES.h and CHARACTER.h
+- Added ATTRIBUTES.cpp and CHARACTER.cpp and implemented logic for each class
+
+##[0.1.0] - 2019-11-24
 ### Added
 - Initial project setup and configuration(s)
 
 ---
-[0.1.0] - 2019-11-24 [CURRENT RELEASE]
+[0.2.0] - 2019-11-24 [CURRENT RELEASE]
 
-[0.0.0] - 2019-11-24 [839199a](https://github.com/Justin-Byrne/Player/commit/839199aeea37eeae2b59b27c044be357f9f8f508) Initial commit.
+[0.1.0] - 2019-11-24 [6ebe3b5](https://github.com/Justin-Byrne/Player/commit/6ebe3b52f2a695698ca2a80ea563aece3c2c7302) Initial upload
+
+[0.0.0] - 2019-11-24 [839199a](https://github.com/Justin-Byrne/Player/commit/839199aeea37eeae2b59b27c044be357f9f8f508) Initial commit
 ---
 
 ## Types of changes